+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 #### :house: Internal
 
 - Find `@rescript/runtime` for Rewatch compiler-args call. https://github.com/rescript-lang/rescript-vscode/pull/1125
+- Use `prepareRename` command (when a new enough ReScript version is used) to speed up the `rename` command. https://github.com/rescript-lang/rescript-vscode/pull/1124
 
 ## 1.64.0